movieticket/api/serializers.py
from django.contrib.auth import authenticate
from rest_framework import serializers
from rest_framework_simplejwt.tokens import RefreshToken

class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only = True)
    
    def validate(self, attrs):
        email = attrs.get("email")
        password = attrs.get("password")
        
        user = authenticate(username=email, password=password)
        
        if not user:
            raise serializers.ValidationError("Invalid email or password")
        
        if user.role == "":
            raise serializers.ValidationError("Admin user cannot login via API")
        
        if not user.is_active:
            raise serializers.ValidationError("User account is inactivate")
        
        refresh = RefreshToken.for_user(user)
        
        return {
            "access":str(refresh.access_token),
            "refresh":str(refresh),
            "user":{
                "id" : user.id,
                "email":user.email,
                "role":user.role
                
            }
        }